Connected Edit/Save Toggle buttons.
laid down groundwork for delete and create new entries.

All of this is across the general js file and written generically. Call distinction is in the specific hbs file.

All of thses changes are only visible in the Activations page ATM as I have notupdated the hbs pages for the rest yet